Struct rocket_failure::errors::Status

pub struct Status {
    pub code: u16,
    pub reason: &'static str,
}

Structure representing an HTTP status: an integer code and a reason phrase.

Usage

Status classes should rarely be created directly. Instead, an associated constant should be used; one is declared for every status defined in the HTTP standard.

Example

A status of 200 OK can be instantiated via the Ok constant:

use rocket::http::Status;

let ok = Status::Ok;

A status of 404 Not Found can be instantiated via the NotFound constant:

use rocket::http::Status;

let not_found = Status::NotFound;

The code and phrase can be retrieved directly:

use rocket::http::Status;

let not_found = Status::NotFound;

assert_eq!(not_found.code, 404);
assert_eq!(not_found.reason, "Not Found");
assert_eq!(not_found.to_string(), "404 Not Found".to_string());

Fields
code: u16

The HTTP status code associated with this status.

reason: &'static str

The HTTP reason phrase associated with this status.

Implementations

impl Status

pub fn new(code: u16, reason: &'static str) -> Status

Creates a new Status with code and reason. This should be used only to construct non-standard HTTP statuses. Use an associated constant for standard statuses.

Example

Create a custom 299 Somewhat Successful status:

use rocket::http::Status;

let custom = Status::new(299, "Somewhat Successful");
assert_eq!(custom.to_string(), "299 Somewhat Successful".to_string());

pub fn class(&self) -> StatusClass

Returns the class of a given status.

Example

use rocket::http::{Status, StatusClass};

let processing = Status::Processing;
assert_eq!(processing.class(), StatusClass::Informational);

let ok = Status::Ok;
assert_eq!(ok.class(), StatusClass::Success);

let see_other = Status::SeeOther;
assert_eq!(see_other.class(), StatusClass::Redirection);

let not_found = Status::NotFound;
assert_eq!(not_found.class(), StatusClass::ClientError);

let internal_error = Status::InternalServerError;
assert_eq!(internal_error.class(), StatusClass::ServerError);

let custom = Status::new(600, "Bizarre");
assert_eq!(custom.class(), StatusClass::Unknown);

pub const Continue: Status

Status with code 100 and reason Continue .

pub const SwitchingProtocols: Status

Status with code 101 and reason Switching Protocols .

pub const Processing: Status

Status with code 102 and reason Processing .

pub const Ok: Status

Status with code 200 and reason OK .

pub const Created: Status

Status with code 201 and reason Created .

pub const Accepted: Status

Status with code 202 and reason Accepted .

pub const NonAuthoritativeInformation: Status

Status with code 203 and reason Non-Authoritative Information .

pub const NoContent: Status

Status with code 204 and reason No Content .

pub const ResetContent: Status

Status with code 205 and reason Reset Content .

pub const PartialContent: Status

Status with code 206 and reason Partial Content .

pub const MultiStatus: Status

Status with code 207 and reason Multi-Status .

pub const AlreadyReported: Status

Status with code 208 and reason Already Reported .

pub const ImUsed: Status

Status with code 226 and reason IM Used .

pub const MultipleChoices: Status

Status with code 300 and reason Multiple Choices .

pub const MovedPermanently: Status

Status with code 301 and reason Moved Permanently .

pub const Found: Status

Status with code 302 and reason Found .

pub const SeeOther: Status

Status with code 303 and reason See Other .

pub const NotModified: Status

Status with code 304 and reason Not Modified .

pub const UseProxy: Status

Status with code 305 and reason Use Proxy .

pub const TemporaryRedirect: Status

Status with code 307 and reason Temporary Redirect .

pub const PermanentRedirect: Status

Status with code 308 and reason Permanent Redirect .

pub const BadRequest: Status

Status with code 400 and reason Bad Request .

pub const Unauthorized: Status

Status with code 401 and reason Unauthorized .

pub const PaymentRequired: Status

Status with code 402 and reason Payment Required .

pub const Forbidden: Status

Status with code 403 and reason Forbidden .

pub const NotFound: Status

Status with code 404 and reason Not Found .

pub const MethodNotAllowed: Status

Status with code 405 and reason Method Not Allowed .

pub const NotAcceptable: Status

Status with code 406 and reason Not Acceptable .

pub const ProxyAuthenticationRequired: Status

Status with code 407 and reason Proxy Authentication Required .

pub const RequestTimeout: Status

Status with code 408 and reason Request Timeout .

pub const Conflict: Status

Status with code 409 and reason Conflict .

pub const Gone: Status

Status with code 410 and reason Gone .

pub const LengthRequired: Status

Status with code 411 and reason Length Required .

pub const PreconditionFailed: Status

Status with code 412 and reason Precondition Failed .

pub const PayloadTooLarge: Status

Status with code 413 and reason Payload Too Large .

pub const UriTooLong: Status

Status with code 414 and reason URI Too Long .

pub const UnsupportedMediaType: Status

Status with code 415 and reason Unsupported Media Type .

pub const RangeNotSatisfiable: Status

Status with code 416 and reason Range Not Satisfiable .

pub const ExpectationFailed: Status

Status with code 417 and reason Expectation Failed .

pub const ImATeapot: Status

Status with code 418 and reason I'm a teapot .

pub const MisdirectedRequest: Status

Status with code 421 and reason Misdirected Request .

pub const UnprocessableEntity: Status

Status with code 422 and reason Unprocessable Entity .

pub const Locked: Status

Status with code 423 and reason Locked .

pub const FailedDependency: Status

Status with code 424 and reason Failed Dependency .

pub const UpgradeRequired: Status

Status with code 426 and reason Upgrade Required .

pub const PreconditionRequired: Status

Status with code 428 and reason Precondition Required .

pub const TooManyRequests: Status

Status with code 429 and reason Too Many Requests .

pub const RequestHeaderFieldsTooLarge: Status

Status with code 431 and reason Request Header Fields Too Large .

pub const UnavailableForLegalReasons: Status

Status with code 451 and reason Unavailable For Legal Reasons .

pub const InternalServerError: Status

Status with code 500 and reason Internal Server Error .

pub const NotImplemented: Status

Status with code 501 and reason Not Implemented .

pub const BadGateway: Status

Status with code 502 and reason Bad Gateway .

pub const ServiceUnavailable: Status

Status with code 503 and reason Service Unavailable .

pub const GatewayTimeout: Status

Status with code 504 and reason Gateway Timeout .

pub const HttpVersionNotSupported: Status

Status with code 505 and reason HTTP Version Not Supported .

pub const VariantAlsoNegotiates: Status

Status with code 506 and reason Variant Also Negotiates .

pub const InsufficientStorage: Status

Status with code 507 and reason Insufficient Storage .

pub const LoopDetected: Status

Status with code 508 and reason Loop Detected .

pub const NotExtended: Status

Status with code 510 and reason Not Extended .

pub const NetworkAuthenticationRequired: Status

Status with code 511 and reason Network Authentication Required .

pub fn from_code(code: u16) -> Option<Status>

Returns a Status given a standard status code code. If code is not a known status code, None is returned.

Example

Create a Status from a known code:

use rocket::http::Status;

let not_found = Status::from_code(404);
assert_eq!(not_found, Some(Status::NotFound));

Create a Status from an unknown code:

use rocket::http::Status;

let not_found = Status::from_code(600);
assert!(not_found.is_none());